requirements.txt
# Core dependencies for 3D reconstruction pipeline
numpy>=1.21.0
opencv-python>=4.5.0
open3d>=0.15.0
scipy>=1.7.0
pyyaml>=6.0

# Intel RealSense SDK
pyrealsense2>=2.50.0

# Visualization (optional)
matplotlib>=3.4.0

# Realman Robot SDK (install from Realman official source)
# pip install Robotic_Arm
